Promote JDBC CONNECTION_INFO_EXTRACTOR to a named singleton class
Turns the inline lambda into ConnectionInfoExtractor (a proper TagExtractor<DBInfo>
with a private-ctor INSTANCE singleton), mirroring PeerConnectionExtractor. Gives
it a name at the call site (span.setTags(dbInfo, ConnectionInfoExtractor.INSTANCE))
and makes it composable. The setTagIfPresent helper moves onto the class.

Verified auto-injected as a transitive helper (JDBCInstrumentationV0Test: 87/87
green) — no manual helperClassNames() entry needed.
